refactor: 비사용 api 삭제
